Extract: assertType into own file
