+++ lib/create-matcher.js
@@ -11,6 +11,8 @@ var stringProto = require("@sinonjs/commons").prototypes.string;
 var typeOf = require("@sinonjs/commons").typeOf;
 var valueToString = require("@sinonjs/commons").valueToString;
 
+var assertType = require("./create-matcher/assert-type");
+
 var arrayIndexOf = arrayProto.indexOf;
 var arrayEvery = arrayProto.every;
 var join = arrayProto.join;
@@ -38,30 +40,6 @@ var matcher = {
  */
 function isMatcher(object) {
     return isPrototypeOf(matcher, object);
-}
-
-/**
- * Ensures that value is of type
- *
- * @private
- * @param {*} value A value to examine
- * @param {string} type A basic JavaScript type to compare to, e.g. "object", "string"
- * @param {string} name A string to use for the error message
- * @throws {TypeError} If value is not of the expected type
- * @returns {undefined}
- */
-function assertType(value, type, name) {
-    var actual = typeOf(value);
-    if (actual !== type) {
-        throw new TypeError(
-            "Expected type of " +
-                name +
-                " to be " +
-                type +
-                ", but was " +
-                actual
-        );
-    }
 }
 
 /**
